+++ app_plus/src/main/java/org/gkisalatiga/plus/services/ClipManager.kt
@@ -1,3 +1,9 @@
+/**
+ * AGPL-3.0-licensed
+ * Copyright (C) GKI Salatiga 2024
+ * Written by Samarthya Lykamanuella (github.com/groaking)
+ */
+
 package org.gkisalatiga.plus.services
 
 import android.content.ClipboardManager